@@ -10,16 +10,16 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~17.3.12",
+    "@angular/animations": "~18.2.8",
     "@angular/cdk": "^13.3.3",
-    "@angular/common": "~17.3.12",
-    "@angular/compiler": "~17.3.12",
-    "@angular/core": "~17.3.12",
-    "@angular/forms": "~17.3.12",
+    "@angular/common": "~18.2.8",
+    "@angular/compiler": "~18.2.8",
+    "@angular/core": "~18.2.8",
+    "@angular/forms": "~18.2.8",
     "@angular/material": "^13.3.3",
-    "@angular/platform-browser": "~17.3.12",
-    "@angular/platform-browser-dynamic": "~17.3.12",
-    "@angular/router": "~17.3.12",
+    "@angular/platform-browser": "~18.2.8",
+    "@angular/platform-browser-dynamic": "~18.2.8",
+    "@angular/router": "~18.2.8",
     "rxjs": "~7.5.0",
     "tslib": "^2.3.0",
     "zone.js": "~0.14.10"
@@ -27,7 +27,7 @@
   "devDependencies": {
     "@angular-devkit/build-angular": "^18.2.8",
     "@angular/cli": "^18.2.8",
-    "@angular/compiler-cli": "~17.3.12",
+    "@angular/compiler-cli": "~18.2.8",
     "@types/jasmine": "~3.10.0",
     "@types/node": "^12.11.1",
     "jasmine-core": "~4.0.0",
